backend/requirements.txt
fastapi
uvicorn
python-dotenv
pyyaml
langgraph
langchain-community
langchain-core
langchain-ollama
chromadb
docker
mcp
pytest
